==> 4_sntiment_analysis.ipynb <==
import pandas as pd
books = pd.read_csv("/books_with_categories.csv")

#lets find a fine tuned model for our task
from transformers import pipeline
classifier = pipeline("text-classification",
                      model="j-hartmann/emotion-english-distilroberta-base",
                      top_k = None,
                      device = -1)
classifier("i am both polite and angry person")

# now we need to decide which level of description we will apply to find the sentiment from the text
# 1 we can find overall sentiment of whole description using classifier
# 2.we can break them into individual sentences and them find

books["description"][0]

we can see that from the description that words in different sentences are totally differernt from one another .So applying classifier here and getting correct sentiment will be bit difficult.
Thats why we break them into discrete parts and then classify them

#see the exampke here that above description is full if mx sentences thats why model is prediction a score of 65% only
classifier(books["description"][0])

#break the description
classifier(books["description"][0].split("."))

sentences = books["description"][0].split(".")
predictions = classifier(sentences)

sentences[0]

predictions[0]

sentences[3]

predictions[3]

sorted(predictions[0],key=lambda x: x["label"])

#lets now find the maximum probability for each description
emotion_labels = ["anger", "disgust", "fear", "joy", "neutral", "sadness", "surprise"]
isbn = []
emotion_score = {label: [] for label in emotion_labels}
def calculate_max_emotion_score(predictions):
  per_emotion_scores = {label: [] for label in emotion_labels}
  for prediction in predictions:
    sorted_predictions = sorted(prediction, key=lambda x: x["label"])
    for index, label in enumerate(emotion_labels):
      per_emotion_scores[label].append(sorted_predictions[index]["score"])
  return {label: max(scores) for label, scores in per_emotion_scores.items()}

for i in range(10):
  isbn.append(books["isbn13"][i])
  sentences = books["description"][i].split(".")
  predictions = classifier(sentences)
  max_scores = calculate_max_emotion_score(predictions)
  for label  in emotion_labels:
    emotion_score[label].append(max_scores[label])

emotion_score

from tqdm import tqdm
emotion_labels = ["anger", "disgust", "fear", "joy", "neutral", "sadness", "surprise"]
isbn = []
emotion_score = {label: [] for label in emotion_labels}
for i in tqdm(range(1000)):
  isbn.append(books["isbn13"][i])
  sentences = books["description"][i].split(".")
  predictions = classifier(sentences)
  max_scores = calculate_max_emotion_score(predictions)
  for label  in emotion_labels:
    emotion_score[label].append(max_scores[label])

emotion_df = pd.DataFrame(emotion_score)
emotion_df["isbn13"] = isbn

emotion_df

books = pd.merge(books,emotion_df,on="isbn13")

books.to_csv("books_with_emotions.csv",index=False)
